Change tokens from set to list for event handling
This small set of changes fixes issues with pywinrt on Python 3.13. Due to changes in the code `tokens.add(tok)` would raise `TypeError: unhashable type: 'winrt._winrt_windows_foundation.EventRegistrationToken'` as it could not accept the `EventRegistrationToken` object. Changing it to a list resolves this. It is both backwards and forwards compatible so older Python versions will not be affected.